Refactors to split fsrs6/fsrs7 logic, add traits to avoid doing if/else, added penalty flag
