+++ docs/MODEL_SPLIT.MD
@@ -1,0 +1,57 @@
+# Model Split
+
+## Purpose
+
+Reduce compatibility risk by separating FSRS-6 and FSRS-7 model-specific math
+while keeping shared model state/config logic in `src/model.rs`.
+
+## File Layout
+
+- Shared entrypoints/types:
+  - `src/model.rs`
+- FSRS-6 specific math:
+  - `src/model_v6.rs`
+- FSRS-7 specific math:
+  - `src/model_v7.rs`
+- FSRS-6 parameter compatibility normalization (`0/17/19 -> 21`):
+  - `src/model_v6.rs::check_and_fill_parameters_fsrs6()`
+- Parameter clipping split by version:
+  - `src/parameter_clipper_v6.rs`
+  - `src/parameter_clipper_v7.rs`
+  - shared dispatcher in `src/parameter_clipper.rs`
+- Inference defaults split by version:
+  - `src/inference_v6.rs` (FSRS-6 defaults/constants)
+  - `src/inference_v7.rs` (FSRS-7 defaults)
+  - shared API/orchestration in `src/inference.rs`
+- Training schedule-penalty split by version:
+  - `src/training_v6.rs` (FSRS-6 no-op penalty path)
+  - `src/training_v7.rs` (FSRS-7 penalty math/gradients)
+  - shared training loop/objective composition in `src/training.rs`
+
+## Dispatch Strategy
+
+- `src/model.rs` keeps shared orchestration (`Model::step`, `Model::forward`).
+- `src/model.rs::check_and_fill_parameters()` keeps shared validation and delegates
+  FSRS-6 compatibility conversion to `model_v6`.
+- Version-specific computations are routed through `VersionOps` trait impls:
+  - `Fsrs6Ops`
+  - `Fsrs7Ops`
+- Model version is selected from parameter count at creation/update time and
+  then reused via `ModelVersion`, instead of repeated runtime checks on `w.len()`.
+- Training penalty routing now also uses `ModelVersion` dispatch, so FSRS-6 and
+  FSRS-7 penalty behavior are separated without checking parameter length in the
+  training loop.
+- Version constants are colocated with version logic:
+  - `src/model_v6.rs`: `PARAM_LEN`
+  - `src/model_v7.rs`: FSRS-7 solver/LUT constants and `PARAM_LEN`
+
+## Data Flow
+
+- Data read:
+  - model parameter vector `w`
+  - runtime tensors (`delta_t`, `rating`, `stability`, `difficulty`)
+- Data updated:
+  - next-step `MemoryStateTensors` during `Model::step()`
+- Data storage:
+  - no new persistence; all values remain in-memory
+  - FSRS-7 S90 LUT cache remains process-local memory only
